+++ coreml-training/iOS App/Sources/Globals.swift
@@ -2,6 +2,3 @@
 
 let labels = Labels()
 let trainingDataset = ImageDataset(split: .train)
-let testingDataset = ImageDataset(split: .test)
-let settings = Settings()
-let history = History.load()
