+++ coreml-training/iOS App/Sources/ImagesByLabel.swift
@@ -1,59 +1,62 @@
 import UIKit
+import DiscreetAI
 
 /**
-  Organizes the images from an ImageDataset grouped by their labels.
-
-  This is used by the "Training Data" and "Test Data" screens, as well as the
-  "Train k-Nearest Neighbors" screen.
+ Organizes the images from an ImageDataset grouped by their labels.
+ 
+ This is used by the "Training Data" and "Test Data" screens, as well as the
+ "Train k-Nearest Neighbors" screen.
  */
 class ImagesByLabel {
-  let dataset: ImageDataset
-  private var groups: [String: [Int]] = [:]
-
-  init(dataset: ImageDataset) {
-    self.dataset = dataset
-    updateGroups()
-  }
-
-  private func updateGroups() {
-    groups = [:]
-    for label in labels.labelNames {
-      groups[label] = dataset.images(withLabel: label)
+    let dataset: ImageDataset
+    let orchestrator: Orchestrator
+    private var groups: [String: [Int]] = [:]
+    
+    init(dataset: ImageDataset, orchestrator: Orchestrator) {
+        self.dataset = dataset
+        self.orchestrator = orchestrator
+        updateGroups()
     }
-  }
-
-  var numberOfLabels: Int { labels.labelNames.count }
-
-  func labelName(of group: Int) -> String { labels.labelNames[group] }
-
-  func numberOfImages(for label: String) -> Int {
-    groups[label]!.count
-  }
-
-  func image(for label: String, at index: Int) -> UIImage? {
-    dataset.image(at: flatIndex(for: label, at: index))
-  }
-
-  func addImage(_ image: UIImage, for label: String) {
-    let filename = UUID().uuidString + ".jpg"
-    let path = "train/\(label)/\(filename)"
-    dataset.addImage(image, for: label, filename: filename)
-    try! orchestrator.addImages(datasetID: "mnist", images: [path], labels: [label])
-    dataset.examples.append((filename, label))
-    // The new image is always added at the end, so we can simply append
-    // the new index to the group for this label.
-    groups[label]!.append(dataset.count - 1)
-  }
-
-  func removeImage(for label: String, at index: Int) {
-    dataset.removeImage(at: flatIndex(for: label, at: index))
-    try! orchestrator.removeImage(datasetID: "mnist", index: index)
-    // All the image indices following the deleted image are now off by one,
-    // so recompute all the groups.
-    updateGroups()
-  }
-
-  func flatIndex(for label: String, at index: Int) -> Int {
-    groups[label]![index]
-  }
+    
+    private func updateGroups() {
+        groups = [:]
+        for label in labels.labelNames {
+            groups[label] = dataset.images(withLabel: label)
+        }
+    }
+    
+    var numberOfLabels: Int { labels.labelNames.count }
+    
+    func labelName(of group: Int) -> String { labels.labelNames[group] }
+    
+    func numberOfImages(for label: String) -> Int {
+        groups[label]!.count
+    }
+    
+    func image(for label: String, at index: Int) -> UIImage? {
+        dataset.image(at: flatIndex(for: label, at: index))
+    }
+    
+    func addImage(_ image: UIImage, for label: String) {
+        let filename = UUID().uuidString + ".jpg"
+        let path = "train/\(label)/\(filename)"
+        dataset.addImage(image, for: label, filename: filename)
+        try! orchestrator.addImages(datasetID: "mnist", images: [path], labels: [label])
+        dataset.examples.append((filename, label))
+        // The new image is always added at the end, so we can simply append
+        // the new index to the group for this label.
+        groups[label]!.append(dataset.count - 1)
+    }
+    
+    func removeImage(for label: String, at index: Int) {
+        dataset.removeImage(at: flatIndex(for: label, at: index))
+        try! orchestrator.removeImage(datasetID: "mnist", index: index)
+        // All the image indices following the deleted image are now off by one,
+        // so recompute all the groups.
+        updateGroups()
+    }
+    
+    func flatIndex(for label: String, at index: Int) -> Int {
+        groups[label]![index]
+    }
 }
