--- coreml-training/iOS App/Sources/SummaryHeaderView.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-class SummaryHeaderView: UIView {
-  @IBOutlet var label: UILabel!
-}
